playwright: attachments in hooks
