+++ packages/allure-js-commons/src/sdk/reporter/ReporterRuntime.ts
@@ -1,4 +1,5 @@
 /* eslint max-lines: 0 */
+import path from "node:path";
 import { extname } from "path";
 import {
   type Attachment,
@@ -130,6 +131,21 @@ export class ShallowStepsStack {
       return undefined;
     };
     return findRecursively(this.steps);
+  }
+
+  addAttachment(attachment: AttachmentOptions, writer: Writer) {
+    const isPath = !!attachment.path;
+    const fileExt = attachment.path ? path.extname(attachment.path) : undefined;
+    const fileName = buildAttachmentFileName({
+      contentType: attachment.contentType,
+      fileExtension: fileExt,
+    });
+    if (isPath) {
+      writer.writeAttachmentFromPath(fileName, attachment.path as string);
+    } else if (attachment.body) {
+      writer.writeAttachment(fileName, attachment.body);
+    }
+    return fileName;
   }
 }
